Added handling of transaction confirmations

+++ wallet/transactions/swaps/bridges/ethereum/ethereum_side.h
@@ -84,9 +84,9 @@ namespace beam::wallet
         bool m_isWithdrawTxSent = false;
         uint64_t m_blockCount = 0;
 
-        uint64_t m_SwapLockTxConfirmations = 0;
+        uint32_t m_SwapLockTxConfirmations = 0;
         uint64_t m_SwapLockTxBlockNumber = 0;
-        uint64_t m_WithdrawTxConfirmations = 0;
+        uint32_t m_WithdrawTxConfirmations = 0;
         uint64_t m_WithdrawTxBlockNumber = 0;
     };
 } // namespace beam::wallet
